Generic module cleanup
- Adds a generic module named TunaModule for any looping class to define certain requirements and start/stop abilities.

- EventManager now extends TunaModule.
